+++ adminforth/types/AdminForthConfig.ts
@@ -1,4 +1,5 @@
 import type { Express } from 'express';
+import type { Writable } from 'stream';
 
 export interface ICodeInjector {
   srcFoldersToSync: Object;
@@ -13,6 +14,7 @@ export interface IConfigValidator {
 export interface IAdminForthHttpResponse {
     setHeader: (key: string, value: string) => void,
     setStatus: (code: number, message: string) => void,
+    blobStream: () => Writable,
 };
 
 /**